64-bit kernel & UEFI bootloader
Modified the Kernel build process to create a 64-bit ELF binary instead of a 32-bit one. Also created a bootloader which works in the UEFI environment. The bootloader can't start the kernel yet, however subsequent work will enable the bootloader to load and start the kernel.

Additionally, the code for the two different routes attempted was added into the archive folder. Future work which is used to figure certain things out will be stored in this folder.